refactor(settings_view): split window chrome + generic controls into chrome.rs
Lifts the structural chrome + reusable form controls out of
mod.rs into a new `settings_view::chrome` child module. About
475 lines covering 11 `pub(super)` helpers:

  * Window structure: `window_header`, `rail`, `scrollable_pane`,
    `section_card_with_desc`, `bool_field`.
  * Generic affordances: `dismiss_notification_after`,
    `encode_file_path`, `resolve_config_dir_display`,
    `import_button`, `trash_button`, `make_select`.

Two pre-existing doc-comment layout bugs surfaced and got
fixed along the way:

  * `dismiss_notification_after`'s doc was duplicated inline
    (6 lines of "Schedule a Notification…" repeated verbatim,
    pre-existing in mod.rs at HEAD). De-duplicated in the move,
    single clean copy now lives above the fn in chrome.rs.

  * `import_button`'s doc had drifted up into the gap between
    `bool_field` and the `SECTION_KEYWORDS` const in HEAD, so
    `import_button` itself rendered undocumented. Reattached
    to the actual fn in chrome.rs.

The taxonomy consts (`SECTION_KEYWORDS`, `TAB_SECTIONS`) stay
in mod.rs because they're data tables consumed by
`SettingsView`'s filter logic rather than chrome primitives —
moving them would force `pub(super)` widenings on
`SettingsTab` enum variants for no functional gain.

mod.rs drops by ~480 lines (3,215 → 2,734).

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
